pyproject.toml
[project]
name = "postgres-fastmcp"
version = "0.0.4"
description = "PostgreSQL Tuning and Analysis Tool"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.0",
    "psycopg[binary]>=3.2.6",
    "humanize>=4.8.0",
    "pglast==7.2.0",
    "attrs>=25.3.0",
    "psycopg-pool>=3.2.6",
    "instructor>=1.7.9",
    "fastmcp>=2.13.0.2",
    "pydantic-settings>=2.11.0",
    "pydantic>=2.12.4",
    "key-value",
    "redis>=5.0.0",
]


[[project.authors]]
name = "Andrey Shlyapin"
email = "andrey@skzd.ru"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project.scripts]
postgres-fastmcp = "postgres_fastmcp.main:main"

[[tool.uv.index]]
url = "https://repo.mng.sbercloud.tech/repository/pypi/simple"

[[tool.uv.index]]
url = "https://repo.mng.sbercloud.tech/repository/cpe_automation/simple"

[dependency-groups]
dev = [
    "devtools>=0.12.2",
    "mypy>=1.17.1",
    "ruff>=0.12.9",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "docker>=7.0.0",
]

[tool.mypy]
plugins = "pydantic.mypy"
python_version = "3.12"
strict = true
exclude = [".venv", "tests"]

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"
default = true

[[tool.mypy.overrides]]
module = [
    "pglast",
    "pglast.*",
    "psycopg",
    "psycopg_pool",
    "humanize",
    "instructor",
    "mcp",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "tests.*",
    "tests",
]

follow_imports = "skip"
ignore_missing_imports = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false

[tool.ruff]
target-version = "py312"
line-length = 120
exclude = [".venv*", "tests"]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D100",  # Missing docstring in public module
    "D104",  # Missing docstring in public package
    "D105",  # Missing docstring in magic method
    "D107",  # Missing docstring in `__init__`
    "RUF001",  # String contains ambiguous unicode character
    "RUF002",  # Docstring contains ambiguous unicode character
    "RUF003",  # Comment contains ambiguous unicode character
    "COM812",  # Trailing comma missing
    "CPY001",  # Copyright ignore
    "BLE001",  # Blind exception
    "PLR1702",
    "PLR0912",
    "PLR0914",
    "PLR0915",
    "TRY400",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    "S101",  # Checks for uses of the `assert` keyword
    # Missing type annotations - not critical for tests
    "ANN001",  # Missing type annotation for function argument
    "ANN002",  # Missing type annotation for `*args`
    "ANN003",  # Missing type annotation for `**kwargs`
    "ANN201",  # Missing return type annotation for public function
    "ANN202",  # Missing return type annotation for private function
    "ANN204",  # Missing return type annotation for magic method
    # Missing docstrings - not critical for tests
    "D100",  # Missing docstring in public module
    "D103",  # Missing docstring in public function
    "D104",  # Missing docstring in public package
    "D105",  # Missing docstring in magic method
    "D107",  # Missing docstring in `__init__`
    "D415",  # Missing blank line before closing quote
    "G004",  # Logging statement uses f-string - not critical
    "SLF001",  # Private member accessed - normal for tests
    "PGH003",  # pytest fixture without explicit return type - normal
    "PT011",  # pytest fixture without explicit return type - normal
    "PLR2004",  # Magic value used in comparison - normal for tests
    "RUF059",  # Use % formatting in logging - not critical
    "PLC0415",  # Import should be from standard library - not critical
    "INP001",  # File is part of an implicit namespace package - not required for tests
    "W293",  # Blank line contains whitespace - will be fixed automatically
    "ASYNC251",  # Async functions should not call time.sleep - acceptable in tests
    "PTH123",  # open() should be replaced by Path.open() - not critical for tests
    "C901",  # Function is too complex - acceptable for integration tests
    "E501",  # Line too long - will be fixed by formatter
    "D200",  # One-line docstring should fit on one line
    "D205",  # 1 blank line required between summary line and description
    "D209",  # Multi-line docstring closing quotes should be on a separate line
    "D212",  # Multi-line docstring summary should start at the first line
    "D101",  # Missing docstring in public class
    "TRY300",  # Consider moving this statement to an `else` block
    "S110",  # `try`-`except`-`pass` detected - acceptable in tests
    "ASYNC230",  # Async functions should not open files - acceptable in tests
    "PTH110",  # os.path.exists() should be replaced by Path.exists() - not critical
    "PTH107",  # os.remove() should be replaced by Path.unlink() - not critical
    "RET503",  # Missing explicit return - acceptable for decorators
    "RET504",  # Unnecessary assignment before return - will check manually
    "RET505",  # Unnecessary else after return - acceptable
    "UP035",  # Import from collections.abc - will be fixed automatically
    "PT022",  # No teardown in fixture - acceptable
    "PT018",  # Assertion should be broken down - acceptable for tests
    "PT006",  # Wrong type passed to pytest.mark.parametrize - acceptable
    "F841",  # Local variable assigned but never used - acceptable in tests
    "W291",  # Trailing whitespace - will be fixed automatically
    "SIM117",  # Use single with statement - acceptable for tests
    "TRY002",  # Create your own exception - not critical for tests
    "TRY003",  # Avoid specifying long messages - not critical for tests
    "EM101",  # Exception must not use string literal - not critical for tests
    "S608",  # Possible SQL injection - acceptable in tests (testing SQL)
    "S105",  # Possible hardcoded password - acceptable in tests
    "ARG005",  # Unused lambda argument - acceptable for test callbacks
    "ARG001",  # Unused function argument - acceptable for mocks and fixtures
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
lines-after-imports = 2
force-wrap-aliases = true
combine-as-imports = true
split-on-trailing-comma = false

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true

[tool.ruff.lint.flake8-bandit]
check-typed-exception = true

[tool.ruff.lint.flake8-unused-arguments]
ignore-variadic-names = true

[tool.ruff.analyze]
detect-string-imports = true

[tool.ruff.format]
docstring-code-format = true

[tool.uv]


